Resolved FindBugs Warnings - Category STYLE
git-svn-id: http://pww.svn.pic.philips.com/DI-dev/trunk/Projects/PurAir/Android@4810 b40d9911-a802-4213-86a8-5840401a66a1

+++ PurAir/src/com/philips/cl/di/dev/pa/customviews/AirView.java
@@ -73,7 +73,7 @@ public class AirView extends View {
 		paint.setTextSize(getPxWithRespectToDip(20));
 		paint.setTextAlign(Paint.Align.CENTER);
         String txt = String.valueOf(percent) + "%";
-        float x1 = (canvas.getWidth() / 2);
+        float x1 = (canvas.getWidth() / (float)2);
         float y1 = (fheight / 2) + getPxWithRespectToDip(10);
         canvas.drawText(txt, x1, y1, paint);
         paint.getTextBounds(txt, 0, txt.length(), rect);
